# technical/docs/swagger.yaml
basePath: /
definitions:
  posts.PostCreateUpdate:
    properties:
      author:
        type: string
      content:
        type: string
      title:
        type: string
    required:
    - author
    - content
    - title
    type: object
  posts.PostRead:
    properties:
      author:
        type: string
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
host: "localhost:8000"
info:
  contact: {}
  description: A simple blog API for managing posts
  title: Blog API
  version: "1.0"
paths:
  /posts:
    get:
      consumes:
      - application/json
      description: Get a list of all blog posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/posts.PostRead'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new blog post
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.PostCreateUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.PostRead'
        "400":
          description: Invalid request body or validation error
          schema:
            type: string
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a blog post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid post ID
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get a single blog post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostRead'
        "400":
          description: Invalid post ID
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update an existing blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.PostCreateUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostRead'
        "400":
          description: Invalid post ID or request body
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
      summary: Update a post
      tags:
      - posts
swagger: "2.0"